backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5
pydantic
pydantic-settings
python-multipart
alembic
psycopg2-binary
pytest
pytest-asyncio
httpx
slowapi
redis
pypdf
python-docx
openai
email-validator
python-dotenv
google-cloud-storage
